+++ docs/source/changelog.rst
@@ -29,6 +29,10 @@ Changelog
 * :feature:`16` OU-model to the Optimal Mean Reversion module added.
 * :support:`16` OU-model documentation.
 * :support:`17` Added Licence, ReadMe, and RoadMap
+* :feature:`19` Minimum profit optimization module added (Lin et al. 2006, Puspaningrum et al. 2010): Finding optimal pre-set boundaries for cointegrated pairs trading strategy.
+* :feature:`19` Cointegrated time series simulation module added (Lin et al. 2006): Simulate cointegrated series that follows AR(1) dynamics.
+* :support:`19` Minimum profit optimization documentation for cointegrated pairs trading strategy.
+* :support:`19` Cointegrated time series simulation documentation.
 * :support:`20` Added API Key Validation.
 * :support:`20` Add install documentation and test on OS/Ubuntu/Windows.
 * :feature:`5` ML Based Pairs Selection (Horta, 2020) and Data Importer added.